updated protected route wrapper
